# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GuardianLayer Enterprise API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TransactionStatus =
  (typeof TransactionStatus)[keyof typeof TransactionStatus];

export const TransactionStatus = {
  ALLOWED: "ALLOWED",
  HELD: "HELD",
  BLOCKED: "BLOCKED",
  APPROVED: "APPROVED",
  REJECTED: "REJECTED",
} as const;

export interface Transaction {
  id: number;
  source: string;
  destination: string;
  amount: number;
  currency: string;
  riskScore: number;
  status: TransactionStatus;
  category: string;
  ipAddress?: string;
  country?: string;
  createdAt: string;
}

export interface TransactionList {
  transactions: Transaction[];
  total: number;
}

export interface TransactionScanRequest {
  source: string;
  destination: string;
  amount: number;
  currency: string;
  category?: string;
  ipAddress?: string;
  country?: string;
}

export interface TransactionScanResult {
  transaction: Transaction;
  riskFactors: string[];
}

export type ApprovalResultStatus =
  (typeof ApprovalResultStatus)[keyof typeof ApprovalResultStatus];

export const ApprovalResultStatus = {
  APPROVED: "APPROVED",
  REJECTED: "REJECTED",
} as const;

export interface ApprovalResult {
  id: number;
  status: ApprovalResultStatus;
  approvedAt: string;
}

export interface DashboardStats {
  totalTransactions: number;
  totalBlocked: number;
  totalHeld: number;
  totalAllowed: number;
  averageRiskScore: number;
  blockRate: number;
  activeAlerts: number;
  integrationsOnline: number;
}

export type RiskTimelineDataPointsItem = {
  date: string;
  avgRisk: number;
  transactionCount: number;
  blockedCount: number;
};

export interface RiskTimeline {
  dataPoints: RiskTimelineDataPointsItem[];
}

export type AlertSeverity = (typeof AlertSeverity)[keyof typeof AlertSeverity];

export const AlertSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface Alert {
  id: number;
  title: string;
  message: string;
  severity: AlertSeverity;
  dismissed: boolean;
  createdAt: string;
}

export interface AlertList {
  alerts: Alert[];
  total: number;
}

export type IntegrationStatus =
  (typeof IntegrationStatus)[keyof typeof IntegrationStatus];

export const IntegrationStatus = {
  online: "online",
  offline: "offline",
  degraded: "degraded",
} as const;

export interface Integration {
  id: string;
  name: string;
  provider: string;
  status: IntegrationStatus;
  lastChecked: string;
}

export interface IntegrationList {
  integrations: Integration[];
}

export type ListTransactionsParams = {
  status?: ListTransactionsStatus;
  limit?: number;
  offset?: number;
};

export type ListTransactionsStatus =
  (typeof ListTransactionsStatus)[keyof typeof ListTransactionsStatus];

export const ListTransactionsStatus = {
  ALLOWED: "ALLOWED",
  HELD: "HELD",
  BLOCKED: "BLOCKED",
  APPROVED: "APPROVED",
} as const;

export type GetRiskTimelineParams = {
  days?: number;
};

export type ListAlertsParams = {
  severity?: ListAlertsSeverity;
  limit?: number;
};

export type ListAlertsSeverity =
  (typeof ListAlertsSeverity)[keyof typeof ListAlertsSeverity];

export const ListAlertsSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;