# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GuardianLayer Enterprise API
 * OpenAPI spec version: 0.1.0
 */
import type { AlertSeverity } from "./alertSeverity";

export interface Alert {
  id: number;
  title: string;
  message: string;
  severity: AlertSeverity;
  dismissed: boolean;
  createdAt: Date;
}