# lib/api-zod/src/generated/types/integration.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GuardianLayer Enterprise API
 * OpenAPI spec version: 0.1.0
 */
import type { IntegrationStatus } from "./integrationStatus";

export interface Integration {
  id: string;
  name: string;
  provider: string;
  status: IntegrationStatus;
  lastChecked: Date;
}