# lib/api-zod/src/generated/types/integrationList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GuardianLayer Enterprise API
 * OpenAPI spec version: 0.1.0
 */
import type { Integration } from "./integration";

export interface IntegrationList {
  integrations: Integration[];
}