# lib/api-zod/src/generated/types/transactionScanRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GuardianLayer Enterprise API
 * OpenAPI spec version: 0.1.0
 */

export interface TransactionScanRequest {
  source: string;
  destination: string;
  amount: number;
  currency: string;
  category?: string;
  ipAddress?: string;
  country?: string;
}