# lib/api-zod/src/generated/types/transactionScanResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GuardianLayer Enterprise API
 * OpenAPI spec version: 0.1.0
 */
import type { Transaction } from "./transaction";

export interface TransactionScanResult {
  transaction: Transaction;
  riskFactors: string[];
}